Revert differences in ORCA output file
There used was a difference in the 'log_cnt' between ORCA and the Postgres
planner, but I'm not seeing that difference in actual behaviour anymore.
I'm not sure what caused it, but surely the upstream behaviour is correct
in any case, even if the other behaviour was also correct. So just
memorize the expected output, by removing the ORCA-specific expected
output file - the behaviour is the same now.

The reset of the reverted differences were cosmetic, but better to keep
the files as close to each other as possible.