Partially revert "Add missing field to read/out-functions."
This reverts the changes to the 'fast' out/read functions from
commit 0b79b0f1361979396cf0f28c8acd5ba7b22e1fac.

On closer look, 'pseudocols' is only used in the planner stage, so there's
no need to have out/read support for it. There's a comment in the struct
explaining that. In master, I'll add a comment on that to the read/out
functions as well.