Allow a temporary object to be schema-qualified with the temp schema.
I think some of the GPDB internal stuff expands the temp schema's name
in the CreateStmt, so let's cope with it.

At least 'join' regression test tripped over this.

I reverted this a few days ago, thinking that it's not needed anymore
for some reason, but I'm now seeing a lot of failures again. I messed
up my testing somehow. In any case, this is clearly still needed, so
revert the revert..