Revert noop policy copy operations
The GpPolicyCopy() function is (for food reasons) a noop when passed
an empty policy so neither of these calls were accomplishing much.
The call in ExecuteQuery() was before the merge inverted and did a
copy in case the policy was asserted to not be null but it was
relaxed to this without causing breakage so keep this until proven
to cause issues and save a memory operation.