Delete commented out calls to set t_tableOid
The t_tableOid struct member, which in Greenplum has been moved to
the TupleTableSlot, was previously set in these cases but there is
no clear way (or apparent reason) to reach the corresponding slot
struct member here. Also, not following upstream and setting it is
making the tests happy so remove the commented out code.