Revert this assertion/elog like it is in the upstream.
In the upstream, this is an assertion, but in GPDB, we've downgraded it to
an elog(LOG). I don't know *why* we've downgraded it, but the comment that
the GPDB_83_MERGE_FIXME complained about actually seems correct to me.

However, since we're now merging the upstream code that introduced this
assertion, let's try to fix this properly, and avoid the unnecessary diff
from upstream going forward, and turn this back into an assertion.