Temporarily revert the "proper" fix for CVE-2013-0255.
This reverts upstream commit 71627f3d, leaving us with the pre-9.3 band-aid
fix for that security issue (ie. the vulnerability is still plugged). We
will re-apply this, as a separate commit, after the merge commit, to keep
the merge commit as lean as possible, and to make it more clear in the git
history that we have cherry-picked the proper fix.

The purpose of this back-and-forth is to have a commit in the git history
that introduces the "proper" fix, with the commit message that we want
to appear in the final, rewritten git history. That makes life easier for
the script that we will use to reconstruct the merge commit and git
history.