Update default value of optimizer_force_XXX gucs
In 2015, to avoid generating plans single stage aggregations due to
cardinality misestimation we force the generation of multi-stage
aggregation even when such a plan is sub-optimal. Since then we have
fixed several of the cardinality estimation bugs. This PR reverts the
default to be based on cost rather than a hint.

In cases, when the cardinality is still wrong user will have the
oppurtunity to force it.